python_projects/cub200_image_segmentation_streamlit/streamlit_app/requirements.txt
albumentations==1.4.14
numpy==1.26.4
opencv-contrib-python==4.10.0.84
pillow==10.4.0
requests==2.32.3
streamlit==1.38.0
torch==2.4.0
torchvision==0.19.0